profiler: merge extra profile after computing delta
For C allocation profiling, the C samples shuold be merged with the Go
allocation profile. Previously this was done by passing the C profile
through to the Delta.Convert function. However, we don't actually want
to compute a delta for the C samples, only the allocations. The C
samples need to be merged after the delta profile is computed.