all: use depaware to track core package dependencies
The github.com/tailscale/depware tool lists the dependencies for an
individual package. depaware also identifies OS-specific dependencies
and dependencies which use unsafe or cgo code. depaware produces a file
which is checked in alongside each package and must be kept up-to-date.
This makes it easy to examine dependency changes during code review.

Add depaware files for the core non-contrib packages, as well as a
GitHub workflow to check the depaware.txt files, a script to update the
files, and documentation for contributors.