profiler/internal/cmemprof: clean up function names
The profile_allocation* functions shouldn't show up in the profile. If
they show up, they'll be leaf nodes, and the profile is meant to show
allocation calls in the leaf nodes, not profiling calls. Also, on Linux
the allocation functions get wrapped, and their names start with a
"__wrap_" prefix. This could be confusing, and I think it's worth the
slight lie of removing the prefix to avoid confusion.

Also avoid creating duplication profile.Functions.