profiler: correctly set agent telemetry proxy URL
The profiler sets agentURL to http://host:port/profiling/v1/input.
So doing path.Join on that to add the telemetry proxy endpoint is
wrong for two reasons: The "//" at the beginning will be converted
to a "/", and the endpoint will be wrong. We need to parse the
URL and replace the path with the telemetry proxy endpoint.