contrib: support tracing for GraphQL
This change adds basic support for tracing
GraphQL functions that use github.com/99designs/gqlgen.

Query and mutation operations can be traced, with child
spans generated for reading, parsing, and validating
operation strings.

Support for subscriptions is limited to mutations that
occur within the context of a subscription, since
many subscription operations are long running (e.g. a trace
would remain open the entire time a client is subscribed to
an endpoint).

Support for field operations is not supported.

Updates #507