profiler: collect metrics, goroutine wait profiles at end of profiling period (#1382)
After moving to collecting profiles concurrently, our intention was to
preserve the previous behavior where all profiles were collected at the
end of the profiling period so that they all cover the same time span of
program execution. We missed the metrics profile (which collected at the
beginning and sometimes overlapped with uploading, see PR #1377) and
the goroutine wait profile, so this commit adds (interruptible) sleeps
to collect those profiles at the end of the profiling period.