profiler/internal/cmemprof: make C.malloc safe to profile
This commit adds workarounds to make C.malloc safe to profile.
Essentially, we replace C.malloc calls with our own idential version
which includes safety checks that C.malloc lacked.

Previously, this function was *not* safe to profile because our sample
recording function called back into Go to record the call stack.  By
calling from C back into Go, our recording function could cause the Go
call stack to grow. Go generates a special Go wrapper for malloc by hand
that lacks a safety check for stack growth which is normally included in
C function wrappers. If we cause the stack to grow during C.malloc, the
return address for malloc will be written to the previous stack
location, which causes the program to panic.

The way we overcome this limitation works like this:

* For any C function called from Go, the Go compiler generates a wrapper
  which calls the C function.
* This wrapper is called through a function pointer. This function
  pointer will have a named symbol that we can locate in the program's
  symbol table (provided the program hasn't been stripped).
* We generate our own, functionally identical malloc wrapper.
* We find the address of our wrapper in the symbol table. Then we
  find the addresses of the C.malloc function pointers, and write
  the address of our wrapper.

The details of this are slightly different between darwin (macOS) and
Linux. In particular, darwin gives us direct access to all of the
program's load commands, which we can access at runtime. On Linux, the
function pointer symbols don't have global scope/external visibility, so
they won't be visible in the dynamic symbol table we would find by
calling dl_iterate_phdr. This means we have to read the executable
ourselves and locate the appropriate sections.

We add a global flag that indicates whether we succesfully found the
C.malloc function pointers and replaced them with our wrappers. If we
*didn't*, e.g. because the program has been stripped, we fall back to
the existing safety check.