example: update example to be usable by default
The previous example would not work by default if
copied into a fresh main() function since it was
reading a test.json file that wouldn't exist.

The new example uses context to create child spans,
which is a common use case for connecting spans. It
will also run by default if someone has the datadog
agent setup and running locally.