profiler/internal/cmemprof: clean up mapping parser sort mappings
The mapping parser originally came from the Go pprof internals, but it
could be a little simpler with the help of a few convenience functions.
Additionally, in cases where a program links a lot of shared libraries
(dozens or more) it can be a tad faster to have the mappings sorted. For
fewer mappings, doing a binary search is somewhat slower, but the
difference is on the order of nanoseconds.