profiler/internal/cmemprof: use Go /proc/self/maps parser (again)
The new implementation was wrong (didn't handle filenames with spaces
correclty). Rather than make it right, just use the Go runtime/pprof
parser instead like I was originally going to do. Add a test case so
that we catch any regressions if somebody tries to refactor the code and
makes the same mistake.

To comply with the conditions of the Go license, include a copy of the
license together with the copyright notice. To comply with Datadog
policy, include a reference to the license in LICENE-3rdparty.csv.