profiler/internal/cmemprof: add full mappings to the profile (#1402)
* profiler/internal/cmemprof: add full mappings to the profile

Since the C allocation profile might contain samples from shared
libraries, the profile should have appropriate mappings for the
different libraries, rather than a single "default" mapping. In
addition, even if we don't have function & file information for a
sample, we can at least show the library it came from.

The /proc/self/maps parser comes from the Go runtime. To comply
with the conditions of the Go license, include a copy of the license
together with the copyright notice. To comply with Datadog
policy, include a reference to the license in LICENE-3rdparty.csv.

Co-authored-by: Felix Geisendörfer <felix@datadoghq.com>