ddtrace/tracer: add WithPropagation() option to propagate user id over spans (#1387)
This change adds an option - WithPropagation() - that can be fed to SetUser() in order to enable propagation of the user id over spans and services. SetUser() options now modify a configuration struct that is passed to span.setUser() which in turn takes care of setting the span tags according to this configuration.