contrib/internal/httptrace: don't use inet.af/netaddr if Go >= 1.19
The inet.af/netaddr package internally imports a package
(go4.org/unsafe/assume-no-moving-gc) which is designed to panic when
imported for newer Go versions. As of Go 1.19, the inet.af/netaddr has
been added to the Go standard library under the net/netip package. We
can prefer this package for versions of Go >= 1.19, and fall back to
inet.af/netaddr for older Go versions.

The packages have nearly identical APIs. This commit abstracts away
which package is being used, and selects either the standard library
version or the inet.af/netaddr version depending on the go1.19 build
tag.