contrib: support tracing for GraphQL (#1380)
* contrib: support tracing for GraphQL

This change adds basic support for tracing
GraphQL functions that use github.com/99designs/gqlgen.

Query and mutation operations can be traced, with child
spans generated for reading, parsing, and validating
operation strings.

Variables of the query are not a part of the span.

Support for subscriptions is limited to mutations that
occur within the context of a subscription, since
many subscription operations are long running (e.g. a trace
would remain open the entire time a client is subscribed to
an endpoint).

Support for field operations is not supported.

A warning from the docs:
Data obfuscation hasn't been implemented for graphql queries yet, and any sensitive data in the query will be sent to Datadog as the resource name of the span. To ensure no sensitive data is included in your spans, always use parameterized graphql queries with sensitive data in variables.

Updates #507

* remove outdated godoc comment

* Update ddtrace/ext/app_types.go

Co-authored-by: Andrew Glaude <andrew.glaude@datadoghq.com>

* contrib: clean up context usage and errors

* run lint

* go mod tidy

* document and test obfuscation

* change service name to graphql

* change default operation name

Co-authored-by: Andrew Glaude <andrew.glaude@datadoghq.com>