ddtrace/tracer: better trace and span ID randomness (#1472)
This resolves an issue that could cause collisions due to random only
using a 32 bit seed. Even though we're generating a nice
cryptographically secure 64 bit seed it actually is only using 32 bits.
This means that if a customer is running a large number (e.g. > 30k) of
go service instances that generate spans, there is a possibility for a
traceid / spanID collision.

To resolve this we XOR the generated 64 bit number with the span start
time, this keeps the random numbers evenly distributed but reduces the
likelihood of a collision should two instances of a service get the same
random seed.