ddtrace: support tracer.SetUser on a mockspan
tracer.SetUser() was specific to the *tracer.span type and therefore was not working
with the mockup span.
To be able to use tracer.SetUser() on the mockup span type, we propose to implement
its behaviour through the now exported `SetUser()` method of the span struct types.
tracer.SetUser() does a type-assertion to check this method is present and delegates
its call to it. We can now provide a proper implementation for *mockspan.