profiler: ensure that CPU profile records profiler work (#1485)
Our profiler library does post-processing of profiles, such as computing
deltas for heap, block, and mutex profiles, which has non-zero CPU
overhead. Right now there's no guarantee that that work will be captured
in the CPU profile. Add synchronization so that the CPU profiler will
only be stopped after all other profiles are completed.

If the CPU profile duration is less than the profiling period, make sure
that the CPU profile happens at the end rather than the beginning of
the profile period. That way, we respect the shorter CPU duration but
still capture the profiler library work.