{contrib/google.golang.org/grpc, ddtrace/tracer}: various test fixes.
contrib/google.golang.org/grpc contained a test that was flaky due to
assuming the order of spans in a slice. That has been fixed.

ddtrace/tracer contained a test that was flaky on Windows due to Windows
timer issues. This test has been modified to give an appropriate leeway
on Windows so that the test will pass.