fastdelta: make more resilient against malformed inputs
Add minimal fuzz testing, and address several issues surfaced by the
fuzzer:

	* Bound check string table indices
	* Bound check when accessing the list of included strings
	* Bound check the location index
	* Check that packed varints are valid, and not too large
	* CHeck the number of values per sample

Any of these issues could have been identified _without_ the fuzzer by
thinking about bounds for every access, but the fuzzer surfaced all of
them quickly and we now have repeatable bad inputs to test against.

The motiviation for this is to have some assurance that the fast
delta computer won't fail dramatically if the underlying input changes
in some unexpected way, or is invalid due to a Go runtime bug.

Still TODO: add a top-level recovery mechanism for handling errors or
panics, so we don't leave the delta computer in a half-updated state.