profiler/internal/fastdelta: hash location addresses correctly
Order matters for hashing the addresses in a call stack. If the sample
has repeated single-value location IDs (rather than packed), we
currently hash them one at a time and append the hash to the hashes
slice, then sort it. But this gets rid of the order the addresses
appear.

To be fully robust, we should collect all the addresses first, and then
hash them after we've visited every location.