Export TracedConn, as we need it when using the db.Conn Raw() API call.
We would prefer to be able to type assert to this interface type, but we cannot type
assert without the type being exported. The alternative would be to use reflection here.

In addition, we want to be able to, on occasion, access the underlying connection object, hence
the addition of the WrappedConn() function.

Signed-off-by: Alec Warner <alec.warner@panther.io>