profiler: don't add an env tag if the user didn't configure env (#1560)
If the user does not supply an "env" value to the client, we currently
send "env:none". Because the client tag gets priority, users will see
"env:none" on their uploaded profiles when they don't configure that
value for the client, even if the agent had its own default env tag.
Instead, we should not send an "env" tag if the user didn't configure
it.