simplify: remove includeMapping tracking
Pruning mappings from the heap profiles is really not a big concern. In
fact I can't even think about a case where there'd be more than one
mapping for the memory profile ... this is only a thing for CPU profiles
and cgo.

This also made throughput go from 106 MB/s to 113 MB/s