profiler/internal/fastdelta: add real heap profile stress test (#1557)
Add a test that repeatedly generates real heap profiles, each time
creating more garbage, and checking the fidelity of the delta against
the original implementation.

The garbage is created using mutual recursion so that there are many new
unique call stacks for each iteration. This is inspired by seeing
sorting in a profile where left and right recursive merging lead to an
exponential number of unique call stacks.

The test is too resource intensive to run on CI. Use an env var to
enable it, so we can run it locally.