go.mod: fix fastdelta molecule dependency
Remove the molecule replace directive and depend on the latest
(unreleased) version instead.

The replace directive does not apply for upstream consumers, so this
patch fixes the builds for anybody that depends on the main branch of
dd-trace-go.

This problem was introduced in f7fee570a5bea245a4f7c1a3d051d918be01af66
by accident.