internal/test-apps: add cgo profiling test app
This app is intended to exercise cgo-related profiling features, such as
C call stack unwinding and C allocation profiling. The app uses a very
popular SQLite library which provides a Go interface on top of the C
library. It also pulls in my work-in-progress C call stack unwinding
library. The run.sh script uses the Vegeta load generator to exercise
the app.