contrib/dimfeld/httptreemux/v5: fix flaky test
During review of the renamer change we discovered that the test could
fail intermittently depending on the random iteration order of
looping over items in the lookup params map.

We now first check to replace parameters surrounded by a set of '/',
i.e. '.../:param/...', and then move on to try and replace the
parameter at the end of the path without a trailing '/', i.e.
'../:param'.

This fix worked even if we ran the test multiple times.