tracer: support `b3multi` alias for `b3` carrier (#1594)
### What does this PR do?

Adds b3multi as a supported propagator for trace context extraction and injection, which will act as a new alias for the existing b3 propagator.

### Motivation

This is to align with the OTel specification.

### Describe how to test/QA your changes

Set `DD_PROPAGATION_STYLE_INJECT` and `DD_PROPAGATION_STYLE_EXTRACT` to `"b3"` and ensure that any headers headers propagate as expected. Then do the same for setting it to `"b3multi"` and verify that the behavior didn't change.