ddtrace/tracer: fix UDS/http url configuration and error messages
This commit cleans up the configuration of the agent URL so that it
reflects the actual URL being used by the Tracer. Previously it would be
left as localhost:8162 if the tracer was using UDS, and the error
messages would mention localhost:8126 even though it was sending to the
agent over unix sockets.

In addition, it cleans up/fixes a few tests associated with this change.