contrib/labstack/{echo, echo.v4}: improve error detection (#1000)
This adds improved error detection to the echo integrations. Previously,
any returned error was recorded as an error in the echo span. This patch
causes the integration to inspect the returned error to determine if it is
a echo.HTTPError which it can use to discriminate between real errors (5xx
and up) or non-errors (4xx and below)

Fixes #987