ci: increase multi-OS test timeout from 10 minutes to 15 minutes
The ddtrace/tracer tests take longer than 10 minutes to run on the
GitHub Windows runners. Since 10 minutes is the default Go test timeout,
that means the tests are now failing to complete. Increase the timeout
to 15 minutes to give more headroom until we can figure out why the
tests are taking so long.