ddtrace/tracer: clean up resources in tests
Tests have been failing to clean up after themselves and causing memory
usage issues during unit test runs.

One cause of this is failing to shut down a tracer after starting it,
which is solved simply by defering tracer.Close() for those tests.

Other tests use newConfig() to generate a tracer config, which they use
for various purposes. Unfortunately newConfig() actually creates a statsd
client for the new config, which needs to be Close()'d in order to
release its resources. It doesn't make much sense to need to Close() a
config, or have to Close() something *in* a config, so this problem
requires a more complicated solution. Instead, the config still
calculates the statsd address, but the actual instantiation of the statsd
client is moved into newTracer, where the worker routines and other
things that require cleanup are started.

This unfortunately requires several other pieces of code to keep a new
reference to the statsd client, rather than just to the config as they
did previously.