ddtrace/tracer: use the configured HTTP client during startup logging
During startup logging, the tracer attempts to connect to the trace
agent. However, this is done using the default client and not the one
the user may have configured. This leads to a spurious diagnostic error
when the user connects to the agent over UDS. This commit connects using
the configured client, and adds a test to connect to an "agent" over UDS
during the startup log.

For #1621