profiler: add pool to re-use byte slices for profiles
Add a sync.Pool of byte slices which will contain serialized profiles.
These slices are re-used after a profile is uploaded.

This is mainly motivated by a data race. The fast delta profiler re-uses
a bytes.Buffer for each profile computation. The underlying byte slice
from that buffer is passed on to be uploaded. This is flagged as a race
by the race detector. While delta profile computation and uploading
_shouldn't_ generally happen at the same time, this is still a logical
race.

The simplest fix for this race is to just allocate a new bytes.Buffer
for each delta computation. This commit attempts to address the race
while still avoiding unnecessary allocation.