profiler: return copy of delta profile data to avoid race (#1660)
Fix a data race from the fast delta computation. We re-use
a bytes.Buffer for each delta computation. The underlying byte slice
from that buffer is passed on to be uploaded. This is a data race
since the slice is now read and written concurrently with no
synchronization. This could lead to corruption if the upload
fails several times and is re-tried while the next delta happens.
Make a copy of the profile instead.