ddtrace/tracer: remove some networking from unit tests (#1656)
Starting a new tracer involves some networking. This turns out
to be very slow on Windows. It's particularly noticeable on tests
like TestEnvVars which start a new tracer hundreds of times.
This commit adds an in-memory HTTP server for use in testing,
and uses the server together with a no-op statsd client to
eliminate networking from the slowest test. This shaves off 10
minutes of CI time on Windows. We can clean up networking in
the other tests in this repo in subsequent PRs.