ddtrace/tracer: make sure task is propagated through context
The trace.NewTask function returns a context with a value representing
the task's ID. If you pass that context to another trace.NewTask call,
the new task will be annotated with the ID of its parent task. This lets
you recover the relationship between tasks in the execution trace.
Create the trace task before calling applyPPROFLabels, so we can give
that function the annotated context. This will create the proper
parent/child relationship when a new span is created.

Also, we don't even need anything besides the span ID here. If/when
pprof labels end up in the trace, we can get the info we're already
recording that way. Until then, the span ID combined with correct
parent/child task relationships is enough to recover whatever we want.