ddtrace/tracer: enable trace writer to optionally retry on failure. (#1636)
This pull request implements trace writer retries when there is a
failure. It does so by cloning the payload before sending it. Retries are
optional and are enabled by default when running in AWS Lambda.

Note that this change introduces an extra allocation for every send
attempt. However, it only does so when the tracer is configured for
retries. Therefore, the extra allocation will only occur when run in
lambda. I feel this is a reasonable trade off.