ddtrace/tracer: replace Windows path characters in fake UDS host (#1713)
We still need to give a full URL to the HTTP client, with the "http"
scheme, but we want the URL to somewhat resemble the UDS path. Windows
paths contain colons and back-slashes. Colons in particular cause
trouble because they're usually followed by a port. Replace those
characters with something else.