tracer: support 128-bit propagation of B3 multi header (#1717)
Supports 128-bit trace id propagation in B3 multi headers. Other carriers will be supported in future PRs, and TODOs have been added in the code where it is needed.

Also fixes the issue where a spanContext needed to have an associated span to access and store 128-bit trace ids. It is now a separate field of the spanContext. Now, the _dd.p.tid tag in the Meta map (upper 64 bits of 128-bit trace ID) is populated from the spanContext upon span finish.