ddtrace/tracer: tag spans when execution trace is enabled (#1735)
Adds a tag to spans created when runtime execution tracing is enabled.

This will have no impact on users unless they explicitly enable runtime
execution tracing, which is not on by default.

For our experimental work correlating runtime execution traces with APM
spans, it's difficult to find & retain spans which overlap with execution
trace recording. This is particularly acute now, because we record traces
infrequently due to the overhead. The tag will make instrumented APM
spans more discoverable in the UI.