contrib/go-redis/redis.v8: fix broken parsing of redis commands
The previous code made some assumptions about the format of the commands
that were not correct, namely that they would always contain at least 1
space. This seems to be true for commands generated by the library, but
not necessarily for commands given as text to the Do function, e.g.
client.Do(ctx, "PING").

This results in a panic due to out-of-bounds indexing.

This commit makes the parsing code not assume this format and adds a test
case to ensure it works correctly for commands with zero spaces.