.github/workflows: integrate APM E2E tests in Github Actions CI (#1784)
This PR enables the APM end-to-end (E2E) tests in Github Actions CI.

At the moment, the default test is enabled for all variants, but the
Single Spans test is only implemented for the chi variant. All other
variants will still run the jobs, but are skipped (in system-tests terms
they are expectedly failing).

Once we verify that this works well, we can implement the rest variants
too, in follow-up PRs.

We had several incidents with features not working correctly due to
incorrect assumptions, and bugs, in some component of the whole APM
tracing pipeline (tracer, agent, backend).

We want to enable end-to-end tests for key use-cases to avoid regressions
in the future. See ATI-2419 for more details.