profiler: document/test that Start restarts the profiler
Start will first stop the profiler if it is already running, and the
re-start it with the given configuration. This behaviour is not
documented. It is also not tested--the current test suite passes if
Start is changed to return early when the profiler is already active.
It is possible that there are apps out there which rely on this behavior
(some of our internal apps do), and it would be a breaking change for
these apps to make Start do nothing if the profiler is already started.
Given that there isn't a strong reason to change the behavior otherwise,
we should document it and make sure it's actually tested.

This commit adds a test to lock in this behavior. It also removes the
TestStartStopIdempotent test. This test made no assertions, and at any
rate Start was not idempotent to begin with.