profiler: mark PprofDiff as deprecated
Instead of making it private right away, make it deprecated instead, to
be removed in a later release. This should hopefully give users a chance
to notice that it's going away and stop using it before it's removed.
pkg.go.dev won't show it, and some IDEs and linters will flag use of
deprecated functions.