ci: run pull request workflows for merge groups (#1809)
If we want to enable a merge queue, we should have our CI also run on PR
that get added to the queue. If CI fails, the PR will be removed from
the queue. Right now we require a PR to be up-to-date with main, and
have green CI, before merging into main. This basically replicates that
requirement for merge queues.