profiler: deprecate PprofDiff (#1806)
The PprofDiff function was added to our public API unintentionally as
part of #1511. The function is only meant for internal use. This was
missed during code review, likely due to the size of the change and the
fact that most of our attention was devoted to making sure the core
parts of the change were implemented correctly.

Mark the function as deprecated, to be removed in a later release.
This should hopefully give users a chance to notice that it's going away
and stop using it before it's removed. pkg.go.dev will show that the
the function is deprecated, and some IDEs and linters will flag use of
deprecated functions.

This is an exception to our normal v1 compatibility guarantees. It is
very unlikely that any of our users depend on this function. It serves
no purpose in the context of the profiler's public API, and anybody who
might have a reason to use this function would probably have already
found a way to do what they want using the
github.com/google/pprof/profile package.