profiler: properly show UDS path in startup log
We don't actually use the UNIX path in our upload requests. The UNIX
path is used in the underlying transport, so the client doesn't care
what the host part of the URL is, as long as the path
("/profiling/v1/input") is right. Nonetheless, we should give some
signal that UDS was configured properly.