profiler: make stopping the limited trace collector more efficient
The limited trace collector currently dispatches a goroutine to stop
tracing when the trace size limit has been exceeded. The rationale
was that tracing will deadlock if we attempt to stop it from within the
io.Writer we pass to the tracer. However, we're creating a new goroutine
every time a write call exceeds the limit. There may be many write calls
in between the first time we hit the limit and when tracing is actually
stopped. This means there are potentially many goroutines piled up in
the Stop method waiting for tracing to be done.

To remedy this, and to make the code easier to reason about, use a
channel to signal that the tracing should stop and a boolean to make
sure we only close the channel once. By using a channel, we can unify
all the stopping conditions in a select statement and only stop tracing
in one place. This makes creating a new goroutine to stop tracing
unnecessary.