profiler: add tag to indicate whether execution tracing is enabled
go_execution_trace_enabled indicates whether execution tracing is
enabled, to distinguish between missing a trace because we don't collect
them every profiling cycle from missing a trace because the feature
isn't turned on.

While we're here, add these extra tracing tags to the batch object
that's built during the profiling cycle. We can't safely read the
execution trace config during the uploading, because it can be modified
concurrently by the profiler after #1978. It makes more sense to have
the profiler add all the needed information, which it can do safely, so
that the upload step can just deal with uploading.