WIP: contrib/database/sql: properly annotate DB operations with execution trace tasks
After #336, spans for database operations are created *after* the
operation. This is necessary because some operations may return ErrSkip
if they aren't supported, and creating spans with those errors is very
noisy. But because of this, execution trace tasks associated with those
spans would appear to be only a few microseconds long since they are
created alongside the span and thus only cover the very end of the
operation.

This commit creates execution trace tasks (if the execution tracer is
enabled) before the operations, and provides a way of communicating to
the APM tracer that it doesn't need to create a task for the span. The
APM tracer will still annotate the task with the span ID even if a task
was already created.

Still TODO: add tests