internal: Improve the Iter performance of LockMap on empty maps (#2130)
Turns out doing RLock and RUnlock can still be slow especially when this code is in a HOT path. For most users this map will be empty. Therefore it's worth speeding up that case at the cost of a tiny bit of extra memory and a bit of complexity.

Cherry-picks across #2127