contrib/{internal/httptrace, net/http}: fix memory leak and request closure (cherry-pick #2138) (#2139)
This pull request addresses one major and one minor performance issue.

The major issue is in contrib/net/http, where (*ServeMux).ServeHTTP adds
an element to it's spanOpts slice every time a request is served. This is
both a race, and it grows the slice unbounded resulting in a memory leak,
and a huge amount of CPU time spent copying the slice.

The second, smaller issues is in contrib/internal/httptrace's
HeaderTagsFromRequest. This function created a closure which captured the
request including all of its headers. This may keep the request alive
longer than necessary. Instead we build the list of required headers and
return a function closing over just that list.

This cherry-picks #2138