ddtrace/tracer: add integrations in startup logs (#2115)
Co-authored-by: Andrew Glaude <andrew.glaude@datadoghq.com>

This PR adds integrations to the Go startup logs. This includes information 
about the integrations available to the Go tracer: if it is available, the 
available version, and if the user has instrumented it. If users cannot see 
spans/traces as expected from a given integration, the logs are helpful to 
have on startup to determine if they are instrumenting it properly. 

A future PR should add support for `net/http` and `database/sql`.