contrib/database/sql: use correct context for trace tasks in statements
The context used to prepare a statement and the context used to execute
a statement need not be the same (see
https://pkg.go.dev/database/sql#DB.PrepareContext)

The execution trace task changes added in #2060 incorrectly used the
statement's context from preparation to derive the execution trace task.
We should be using the context provided for executing the statement
instead.

Fixes #2172