ci: run smoke-test with the latest stable Go release (#2209)
The smoke test asserts that our packages still build and pass their unit
tests when all of the dependencies are upgraded to their latest
compatible versions. Sometimes dependencies can drop support for older
Go versions and will no longer compile if we're running the tests using
the unsupported Go version. This can happen, for example, if we are
doing a patch release for a release which supported an older Go version,
and a new Go version comes out. If a user does what this test does, this
they should expect to sometimes need to upgrade their Go toolchain for
this reason. So let's run the test using the latest toolchain.