telemetry: fix profiler not passing config and align data model with spec AIT-7935 (#2248)
The commit fixes 2 bugs

The profiler starts the telemetry client without passing any config so we end up without a valid agent url for telemetry. This is problematic typically when the profiler is used alone or if it starts before the tracer.

The app-client-configuration-change requests contain invalid payloads, i.e used conf_key_values instead of configuration. This seemed to be an inconsistency in the api docs that have been fixed after telemetry was implemented in the tracer. Also took the opportunity to align the remote_config field with the spec.