tracer TestWithHeaderTags: Clear header tags to work with -count=2
This test currently fails when run with -count greater than 2 because
it does not clear all the global state. Add a test to ensure there
are no header tags set at the end. Clear all headers tags after each
test.

This should ensure other tests aren't accidentally depending on this.
It will help get the tests to pass with -count=10, which can be
useful for tracking down flaky tests.