profiler: remove PprofDiff (#2267)
We deprecated PprofDiff in v1.49.0, with a note that it would be removed
in a future release. There is a low likelihood that any user actually
depends on this, and at any rate we have had several releases during
which time staticcheck/other linters could flag the deprecated function.
Finally remove it and inline it into the one place it was used.